refactor: review attachments input lib
