+++ libs/attachment-input/src/constants/attachment-group.ts
@@ -1,0 +1,4 @@
+/** Total attachment count at/above which the grid collapses behind a "+N" tile by default. */
+export const ATTACHMENT_COLLAPSE_THRESHOLD = 5;
+/** Number of tiles shown before the "+N" tile while collapsed. */
+export const ATTACHMENT_COLLAPSED_VISIBLE_COUNT = 4;
