+++ libs/attachment-input/src/constants/upload.ts
@@ -1,1 +1,2 @@
+/** Maximum number of attachment uploads allowed within a rolling one-minute window. */
 export const MAX_UPLOADS_PER_MINUTE = 100;
