+++ libs/attachment-input/src/hooks/useClipboardPaste.ts
@@ -1,7 +1,7 @@
 import type { Attachment } from '@epam/ai-dial-chat-shared';
 import { AttachmentType, RequestStatus } from '@epam/ai-dial-chat-shared';
 import { ClipboardEventHandler, useCallback } from 'react';
-import { generateAttachmentId } from '../utils/generateAttachmentId';
+import { generateAttachmentId } from '../utils/attachment';
 
 /**
  * Returns a `handlePaste` handler for a textarea that intercepts clipboard
