+++ libs/attachment-input/src/index.ts
@@ -9,12 +9,31 @@ export type {
   AttachmentCardProps,
   AttachmentCardColors,
   AttachmentCardTypography,
-} from './models/AttachmentCard';
-export type { AttachmentTrayProps } from './models/AttachmentTray';
-export type { AttachmentGroupProps } from './models/AttachmentGroup';
-export type { AttachmentFileRowProps } from './models/AttachmentFileRow';
-export type { AttachmentMoreTileProps } from './models/AttachmentMoreTile';
-export type { FileDndOverlayProps } from './models/FileDndOverlay';
+  AttachmentCardStyles,
+  AttachmentCardLabels,
+  AttachmentCardState,
+} from './models/attachment-card';
+export type {
+  AttachmentTrayProps,
+  AttachmentTrayLabels,
+} from './models/attachment-tray';
+export {
+  AttachmentTilesLayout,
+  type AttachmentGroupProps,
+  type AttachmentGroupLabels,
+  type AttachmentGroupStyles,
+  type AttachmentTilesPlan,
+} from './models/attachment-group';
+export type {
+  AttachmentFileRowProps,
+  AttachmentFileRowLabels,
+  AttachmentFileRowStyles,
+} from './models/attachment-file-row';
+export type { AttachmentMoreTileProps } from './models/attachment-more-tile';
+export type {
+  FileDndOverlayProps,
+  FileDndOverlayStyles,
+} from './models/file-dnd-overlay';
 
 export { useClipboardPaste } from './hooks/useClipboardPaste';
 export {
@@ -22,24 +41,19 @@ export {
   LazyImageLoadStatus,
 } from './hooks/useLazyImageLoad';
 
-export { generateAttachmentId } from './utils/generateAttachmentId';
 export {
+  generateAttachmentId,
   getAttachmentCardState,
   getExtFromContentType,
-} from './utils/getAttachmentCardState';
-export type { AttachmentCardState } from './utils/getAttachmentCardState';
-export { getAttachmentIcon } from './utils/getAttachmentIcon';
-export {
+  getAttachmentIcon,
   getAttachmentTilesPlan,
-  AttachmentTilesLayout,
-  ATTACHMENT_COLLAPSE_THRESHOLD,
-  ATTACHMENT_COLLAPSED_VISIBLE_COUNT,
-} from './utils/getAttachmentGroupLayout';
-export type { AttachmentTilesPlan } from './utils/getAttachmentGroupLayout';
-export { getNameWithoutExtension } from './utils/getNameWithoutExtension';
-export {
+  getNameWithoutExtension,
   mimeTypesToExtensionLabels,
   isMimeTypeAllowed,
-} from './utils/attachment-mime';
+} from './utils/attachment';
 
+export {
+  ATTACHMENT_COLLAPSE_THRESHOLD,
+  ATTACHMENT_COLLAPSED_VISIBLE_COUNT,
+} from './constants/attachment-group';
 export { MAX_UPLOADS_PER_MINUTE } from './constants/upload';
