+++ libs/attachment-input/src/models/attachment-tray.ts
@@ -1,5 +1,17 @@
 import type { DisplayAttachment } from '@epam/ai-dial-chat-shared';
 
+/** Localised accessible labels for the `AttachmentTray` component. */
+export interface AttachmentTrayLabels {
+  /** Accessible label for the tray region. Defaults to `'Attached files'`. */
+  ariaLabel?: string;
+  /** Accessible label for each card's remove button. */
+  removeLabel?: string;
+  /** Accessible label for each card's retry button (error state only). */
+  retryLabel?: string;
+  /** Accessible label forwarded to each card's root when it is interactive. When omitted, the card's own default (`'Open attachment'`) applies. */
+  clickLabel?: string;
+}
+
 /** Props accepted by the `AttachmentTray` component. */
 export interface AttachmentTrayProps {
   /** The list of attachments to display. */
@@ -10,16 +22,10 @@ export interface AttachmentTrayProps {
   onRetry?: (id: string) => void;
   /** Called when the user clicks a pasted-text card to expand its content back into the input. */
   onExpand?: (id: string) => void;
-  /** Accessible label for the tray region. */
-  ariaLabel?: string;
-  /** Accessible label for each card's remove button. */
-  removeLabel?: string;
-  /** Accessible label for each card's retry button (error state only). */
-  retryLabel?: string;
   /** Called when the user clicks or keyboard-activates an attachment card. Receives the full `DisplayAttachment` object. */
   onAttachmentClick?: (attachment: DisplayAttachment) => void;
-  /** Accessible label forwarded to each card's root when it is interactive. When omitted, the card's own default (`'Open attachment'`) applies. */
-  clickLabel?: string;
+  /** Localised accessible labels for the tray region and each card's interactive elements. */
+  labels?: AttachmentTrayLabels;
   /** Extra class name(s) merged onto the root element. */
   className?: string;
 }
