--- libs/attachment-input/src/utils/getNameWithoutExtension.ts
@@ -1,7 +1,0 @@
-/**
- * Returns the provided name without its trailing file extension.
- */
-export const getNameWithoutExtension = (name: string): string => {
-  const dotIndex = name.lastIndexOf('.');
-  return dotIndex > 0 ? name.slice(0, dotIndex) : name;
-};
